remove fixtures
