+++ client/src/pages/Configurations/ConfigBuild/CustomCodes/CsvImport/ImportCustomCodes.test.tsx
@@ -2,15 +2,12 @@ import { render, screen, within } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 import { TestQueryClientProvider } from '../../../../../test-utils';
 import { ImportCustomCodes } from './ImportCustomCodes';
-import {
-  MOCK_CONFIG_ID,
-  mockCodeSystems,
-  mockIndexedSystem,
-} from '../../../../../utils/fixtures';
 import userEvent from '@testing-library/user-event';
 import { useUploadCustomCodesCsv } from '../../../../../api/configurations/configurations';
 import { Mock } from 'vitest';
 import {
+  DbCodeSystem,
+  IndexedCodeSystem,
   UploadCustomCodesPreviewItem,
   UploadCustomCodesPreviewResponse,
 } from '../../../../../api/schemas';
@@ -271,6 +268,72 @@ function checkCode(codeSystemName: string, exists = true) {
   totalExpectation(`${codeSystemName} Example`);
 }
 
+function csvToDict(csv: string) {
+  const lines = csv.trim().split('\n');
+  const headers = lines[0].split(',').map((h) => h.trim());
+  const indexedValues: Record<string, string>[] = [];
+  lines.slice(1).forEach((line) => {
+    const rowObject: Record<string, string> = {};
+    const values = line.split(',');
+
+    headers.forEach((_, j) => {
+      rowObject[headers[j]] = values[j];
+    });
+    indexedValues.push(rowObject);
+  });
+
+  return indexedValues;
+}
+
+export const mockCodeSystems: DbCodeSystem[] = [
+  {
+    id: '157a00b0-62e6-48c8-b822-475c5d855f3f',
+    key: 'snomed',
+    display_name: 'SNOMED',
+    oid: '2.16.840.1.113883.6.96',
+  },
+  {
+    id: 'bd5ad8fd-f94c-4fcf-97ee-5b63c2e7a42b',
+    oid: '2.16.840.1.113883.6.1',
+    key: 'loinc',
+    display_name: 'LOINC',
+  },
+  {
+    id: '375d4fd5-81f8-4b9e-abd9-979c7987691f',
+    oid: '2.16.840.1.113883.6.90',
+    key: 'icd10',
+    display_name: 'ICD-10',
+  },
+  {
+    id: 'c645801a-26f2-495c-b07f-e9be5ac26275',
+    oid: '2.16.840.1.113883.6.88',
+    key: 'rxnorm',
+    display_name: 'RxNorm',
+  },
+  {
+    id: '4306c91c-a8e2-4f4b-b673-0da9a6432b38',
+    oid: '2.16.840.1.113883.12.292',
+    key: 'cvx',
+    display_name: 'CVX',
+  },
+  {
+    id: 'f65063a3-6836-41ce-8ab8-253994907faa',
+    oid: 'Other',
+    key: 'other',
+    display_name: 'Other',
+  },
+];
+
+export const mockIndexedSystem: IndexedCodeSystem = mockCodeSystems.reduce(
+  (acc: Record<string, DbCodeSystem>, curCodeSystem) => {
+    acc[curCodeSystem.key] = curCodeSystem;
+    return acc;
+  },
+  {}
+);
+
+export const MOCK_CONFIG_ID = 'd8cf3930-a7c2-4761-9ba9-ce72ff9191c8';
+
 const uploadLines = csvToDict(MOCK_UPLOAD_CSV);
 const mockPreviewItems: UploadCustomCodesPreviewItem[] = uploadLines.map(
   (row, i) => {
@@ -290,20 +353,3 @@ const mockUploadResponse: UploadCustomCodesPreviewResponse = {
   total_custom_codes_in_configuration: mockPreviewItems.length,
   code_systems: mockIndexedSystem,
 };
-
-function csvToDict(csv: string) {
-  const lines = csv.trim().split('\n');
-  const headers = lines[0].split(',').map((h) => h.trim());
-  const indexedValues: Record<string, string>[] = [];
-  lines.slice(1).forEach((line) => {
-    const rowObject: Record<string, string> = {};
-    const values = line.split(',');
-
-    headers.forEach((_, j) => {
-      rowObject[headers[j]] = values[j];
-    });
-    indexedValues.push(rowObject);
-  });
-
-  return indexedValues;
-}
